# packages/api-client-ts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/feedback": {
      "post": {
        "operationId": "Feedback_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackCreatedDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "Signed in, but not allowed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The in-app feedback button (students and parents).",
        "tags": [
          "feedback"
        ]
      }
    },
    "/v1/admin/feedback": {
      "get": {
        "operationId": "Feedback_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "NEW",
                "READ",
                "DONE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackListDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "Signed in, but not allowed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Everything sent with the feedback button, newest first.",
        "tags": [
          "feedback"
        ]
      }
    },
    "/v1/admin/feedback/{id}": {
      "patch": {
        "operationId": "Feedback_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackItemDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "Signed in, but not allowed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a message read or done.",
        "tags": [
          "feedback"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kids Coding Platform API",
    "description": "One API for the student and parent web app, the admin panel, the marketing site and the Flutter app. Every route checks the caller’s role.",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateFeedbackDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "SAFETY",
              "BUG",
              "IDEA",
              "PRAISE",
              "OTHER"
            ]
          },
          "message": {
            "type": "string",
            "minLength": 2,
            "maxLength": 1000
          },
          "pagePath": {
            "type": "string",
            "description": "The page it was sent from, e.g. \"/ar/learn/builder-m01-l03\" (query and fragment are dropped).",
            "maxLength": 200,
            "pattern": "^\\/[\\w\\-/[\\].]*$"
          },
          "languageCode": {
            "type": "string",
            "description": "The interface language when it was sent.",
            "pattern": "^[a-z]{2}$"
          }
        },
        "required": [
          "kind",
          "message"
        ]
      },
      "FeedbackCreatedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "FeedbackSenderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "roleKey": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true,
            "description": "Adults: their name. Students: their nickname."
          }
        },
        "required": [
          "id",
          "roleKey",
          "name"
        ]
      },
      "FeedbackItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "enum": [
              "BUG",
              "IDEA",
              "PRAISE",
              "OTHER",
              "SAFETY"
            ],
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "pagePath": {
            "type": "string",
            "nullable": true
          },
          "languageCode": {
            "type": "string"
          },
          "status": {
            "enum": [
              "NEW",
              "READ",
              "DONE"
            ],
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "sender": {
            "nullable": true,
            "description": "Null when the account was deleted.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/FeedbackSenderDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "kind",
          "message",
          "pagePath",
          "languageCode",
          "status",
          "createdAt",
          "sender"
        ]
      },
      "FeedbackListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedbackItemDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "unread": {
            "type": "number",
            "description": "Messages nobody has read yet (for the menu badge)."
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize",
          "unread"
        ]
      },
      "UpdateFeedbackDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "READ",
              "DONE"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}